+++ src/libtcod/sdl2/gl_alias.cpp
@@ -118,8 +118,7 @@ class OpenGLTilesetAlias::impl : public TilesetObserver {
   }
   static pool_type pool_;
  protected:
-  virtual void on_tileset_changed(
-      const std::vector<std::pair<int, Tile&>> &changes) override
+  virtual void on_tileset_changed(const changed_tiles &changes) override
   {
     for (const auto& changed : changes) {
       // If this glyph is already in this alias then free it.